pyproject.toml
[project]
name = "langgraph-course"
version = "0.1.0"
description = "This repository contains modular, hands-on examples exploring key concepts and components in building LLM-powered applications using Python, LangChain, LangGraph and agentic AI workflows."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "gradio>=5.39.0",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.5",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "notebook>=7.4.4",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
